# Fix performance on view graphql_account_rewards