--- a/backend/Application/Database/DatabaseMigrator.cs
+++ b/backend/Application/Database/DatabaseMigrator.cs
@@ -100,6 +100,7 @@ namespace Application.Database
                 .WithScriptsEmbeddedInAssembly(_sqlScriptsAssembly, scriptPath => scriptPath.EndsWith(".sql") && scriptPath.Contains($".{sqlScriptsFolder}."))
                 .LogTo(_dbUpLogWrapper)
                 .WithTransaction()
+                .WithExecutionTimeout(_settings.MigrationTimeout)
                 .Build();
             return upgrader;
         }