--- a/backend/Application/Database/DatabaseSettings.cs
+++ b/backend/Application/Database/DatabaseSettings.cs
@@ -4,4 +4,8 @@ public class DatabaseSettings
 {
     public string ConnectionString { get; init; }
     public string ConnectionStringNodeCache { get; init; }
+    /// <summary>
+    /// Configurable timeout for migrations. Defaults to 5 minutes.
+    /// </summary>
+    public TimeSpan MigrationTimeout { get; init; } = TimeSpan.FromMinutes(5);
 }