--- a/backend/Application/appsettings.json
+++ b/backend/Application/appsettings.json
@@ -1,7 +1,8 @@
 {
   "PostgresDatabase": {
     "ConnectionString" : "Host=postgres;Port=5432;Database=ConcordiumScan;User ID=postgres;Password=passwordFTB2021",
-    "ConnectionStringNodeCache" : "Host=postgres;Port=5432;Database=ConcordiumScan_node_cache;User ID=postgres;Password=passwordFTB2021"
+    "ConnectionStringNodeCache" : "Host=postgres;Port=5432;Database=ConcordiumScan_node_cache;User ID=postgres;Password=passwordFTB2021",
+    "MigrationTimeout": "00:05:00"
   },
   "ConcordiumNodeGrpc": {
     "Address": "http://ccnode:20000"